refactor: 废弃部分类、包调整
