+++ spring-boot-example/spring-authorization-server/server-commons/src/main/java/com/livk/auth/server/common/core/resource/CustomOpaqueTokenIntrospector.java
@@ -18,7 +18,7 @@ package com.livk.auth.server.common.core.resource;
 
 import com.livk.auth.server.common.core.principal.Oauth2User;
 import com.livk.auth.server.common.service.Oauth2UserDetailsService;
-import com.livk.commons.spring.context.SpringContextHolder;
+import com.livk.commons.SpringContextHolder;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.core.Ordered;
