+++ spring-boot-example/spring-redis/redis-repeated-submit/src/main/java/com/livk/redis/submit/support/LockSupport.java
@@ -16,7 +16,7 @@
 
 package com.livk.redis.submit.support;
 
-import com.livk.commons.spring.context.SpringContextHolder;
+import com.livk.commons.SpringContextHolder;
 import com.livk.context.lock.DistributedLock;
 import com.livk.context.lock.LockType;
 import com.livk.context.lock.support.RedissonLock;
