+++ spring-boot-example/spring-redis/redis-repeated-submit/src/main/java/com/livk/redis/submit/support/RedisSupport.java
@@ -16,7 +16,7 @@
 
 package com.livk.redis.submit.support;
 
-import com.livk.commons.spring.context.SpringContextHolder;
+import com.livk.commons.SpringContextHolder;
 import com.livk.context.redis.RedisOps;
 import lombok.experimental.UtilityClass;
 import org.springframework.data.redis.core.ValueOperations;
