+++ spring-boot-extension-tests/curator-spring-boot-test/src/main/java/com/livk/curator/util/ZookeeperUtils.java
@@ -16,7 +16,7 @@
 
 package com.livk.curator.util;
 
-import com.livk.commons.spring.context.SpringContextHolder;
+import com.livk.commons.SpringContextHolder;
 import com.livk.commons.util.ObjectUtils;
 import lombok.experimental.UtilityClass;
 import lombok.extern.slf4j.Slf4j;
