+++ spring-extension-commons/src/main/java/com/livk/commons/Customizer.java
@@ -14,26 +14,21 @@
  * limitations under the License.
  */
 
-package com.livk.commons.spring.context;
-
-import java.lang.annotation.Documented;
-import java.lang.annotation.ElementType;
-import java.lang.annotation.Retention;
-import java.lang.annotation.RetentionPolicy;
-import java.lang.annotation.Target;
-import org.springframework.context.annotation.Import;
+package com.livk.commons;
 
 /**
- * 自动装配相关注解
- * <p>
- * 适用与EnableXXXX
+ * 自定义回调接口
  *
+ * @param <T> the type parameter
  * @author livk
  */
-@Target(ElementType.ANNOTATION_TYPE)
-@Retention(RetentionPolicy.RUNTIME)
-@Documented
-@Import(AutoImportSelector.class)
-public @interface AutoImport {
+@FunctionalInterface
+public interface Customizer<T> {
+
+	/**
+	 * 执行回调过程
+	 * @param t 待执行数据
+	 */
+	void customize(T t);
 
 }
