+++ spring-extension-commons/src/main/java/com/livk/commons/util/ProviderLoader.java
@@ -30,6 +30,7 @@ import java.util.ServiceLoader;
  * @see ServiceLoader
  * @see SpringFactoriesLoader
  */
+@Deprecated(since = "1.3.0", forRemoval = true)
 public sealed interface ProviderLoader permits ProviderLoader.AbstractLoader {
 
 	/**
